# data_aggregator_service/data-aggregator-service-openapi.yml
openapi: 3.0.3
info:
  title: Data Aggregator Service API
  description: |
    Микросервис для агрегации данных о перемещениях объектов, генерации отчетов и аналитики.
    Периодически собирает данные от positioning-service, агрегирует их и предоставляет аналитические отчеты.
  version: 1.0.0
  contact:
    name: Команда разработки
    email: analytics@rtls.example.com
servers:
  - url: http://localhost:8085/api/v1
    description: Локальный сервер разработки
  - url: https://analytics.rtls.example.com/api/v1
    description: Продуктивный сервер
...
paths:
...
  /export/csv:
    get:
      tags:
        - Export
      summary: Экспорт данных в CSV
      description: Экспорт отчета в формате CSV
      operationId: exportToCSV
      parameters:
        - name: report_type
          in: query
          required: true
          schema:
            type: string
            enum: [zone_occupancy, time_in_zone, workflow_efficiency, anomalies]
        - name: start_time
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Файл CSV для скачивания
          content:
            text/csv:
              schema:
                type: string
                format: binary
        '400':
          description: Некорректные параметры запроса
  /export/excel:
    get:
      tags:
        - Export
      summary: Экспорт данных в Excel
      description: Экспорт отчета в формате Excel
      operationId: exportToExcel
      parameters:
        - name: report_type
          in: query
          required: true
          schema:
            type: string
            enum: [zone_occupancy, time_in_zone, workflow_efficiency, anomalies]
        - name: start_time
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Файл Excel для скачивания
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        '400':
          description: Некорректные параметры запроса
  /export/pdf:
    get:
      tags:
        - Export
      summary: Экспорт данных в PDF
      description: Экспорт отчета в формате PDF с визуализацией
      operationId: exportToPDF
      parameters:
        - name: report_type
          in: query
          required: true
          schema:
            type: string
            enum: [zone_occupancy, time_in_zone, workflow_efficiency, anomalies]
        - name: start_time
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: include_charts
          in: query
          required: false
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Файл PDF для скачивания
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          description: Некорректные параметры запроса
...